SOF-2238 No longer possible to delete a PanelLayoutConfiguration if there is a PanelConfiguration using it

+++ src/business-logic/services/panel-service-implementation.ts
@@ -34,6 +34,11 @@ export class PanelServiceImplementation implements PanelService {
   }
 
   public async deletePanelLayoutConfiguration(panelLayoutConfigurationId: string): Promise<void> {
+    const panelConfigurationsForPanelLayoutConfiguration: PanelConfiguration[] = await this.panelConfigurationRepository.getPanelConfigurationsForPanelLayoutConfiguration(panelLayoutConfigurationId)
+    if (panelConfigurationsForPanelLayoutConfiguration.length > 0) {
+      const panelConfigurationIds: string[] = panelConfigurationsForPanelLayoutConfiguration.map(panelConfiguration => panelConfiguration.id)
+      throw new UnsupportedOperationException(`Can't delete PanelLayoutConfiguration for ${panelLayoutConfigurationId}. It's in use by the ConfigurationPanels [${panelConfigurationIds}]`)
+    }
     await this.panelLayoutConfigurationRepository.deletePanelLayoutConfiguration(panelLayoutConfigurationId)
     this.panelEventEmitter.emitPanelLayoutConfigurationDeleted(panelLayoutConfigurationId)
   }
